+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/Rules.kt
@@ -23,7 +23,7 @@ object Rules {
     // sources
     const val IosAppDir = "iosApp"
     const val SrcAndroidMainDir = "*/src/androidMain"
-    const val SrcIosMainDir = "*/src/iosMain"
+    const val SrcIosMainDir = "*/src/ios*"
     const val SrcJsMainDir = "*/src/jsMain"
     const val SrcJvmMainDir = "*/src/jvmMain"
     const val SharedPresentationDir = "shared/presentation"
@@ -60,7 +60,7 @@ object Rules {
     const val AppNavigationModalProvider = "${AppNavigationDir}/ModalProvider.kt"
     const val AppNavigationPermanentProvider = "${AppNavigationDir}/PermanentProvider.kt"
     const val AppNavigationRailProvider = "${AppNavigationDir}/RailProvider.kt"
-    const val ShowcasesDir = "${AppPresentationDir}/showcases"
+    const val ShowcasesDir = "*/presentation/showcases"
     const val ShowcasesDataFlowDir = "${ShowcasesDir}/dataflow"
     const val ShowcasesUserFlowDir = "${ShowcasesDir}/userflow"
     const val ShowcasesHttpDir = "${ShowcasesDataFlowDir}/http"
@@ -68,6 +68,7 @@ object Rules {
     const val ShowcasesPagingDir = "${ShowcasesDataFlowDir}/paging"
     const val ShowcasesKeyValueDir = "${ShowcasesDataFlowDir}/keyvalue"
     const val ShowcasesSqlDelightDir = "${ShowcasesDataFlowDir}/sqldelight"
+    const val ShowcasesRoomDir = "${ShowcasesDataFlowDir}/room"
     const val ShowcasesNavigationDir = "${ShowcasesUserFlowDir}/navigation"
     const val ShowcasesThemeDir = "${ShowcasesUserFlowDir}/theme"
     const val ShowcasesLoaderDir = "${ShowcasesUserFlowDir}/loader"
