+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/dataflow/keyvalue/common/CommonKeyValueProcessor.kt
@@ -6,7 +6,7 @@ import kotli.engine.TemplateState
 import kotli.engine.template.rule.RemoveFile
 import kotli.engine.template.rule.RemoveMarkedLine
 import kotli.template.multiplatform.compose.Rules
-import kotli.template.multiplatform.compose.showcases.datasource.keyvalue.KeyValueShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.dataflow.keyvalue.KeyValueShowcasesProcessor
 import kotlin.time.Duration.Companion.minutes
 
 object CommonKeyValueProcessor : BaseFeatureProcessor() {
