+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/essentials/navigation/NavigationProvider.kt
@@ -3,7 +3,7 @@ package kotli.template.multiplatform.compose.essentials.navigation
 import kotli.engine.FeatureProcessor
 import kotli.template.multiplatform.compose.essentials.EssentialsProvider
 import kotli.template.multiplatform.compose.essentials.navigation.jetpack.JetpackProcessor
-import kotli.template.multiplatform.compose.showcases.navigation.NavigationShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.navigation.NavigationShowcasesProcessor
 
 object NavigationProvider : EssentialsProvider() {
