+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/showcases/ShowcasesProvider.kt
@@ -4,16 +4,16 @@ import kotli.engine.BaseFeatureProvider
 import kotli.engine.FeatureProcessor
 import kotli.engine.FeatureType
 import kotli.engine.model.FeatureTypes
-import kotli.template.multiplatform.compose.showcases.datasource.cache.CacheShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.datasource.http.HttpShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.datasource.keyvalue.KeyValueShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.datasource.paging.PagingShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.feature.loader.data.DataLoaderShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.feature.passcode.PasscodeShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.feature.theme.ThemeShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.feature.theme.change.ChangeThemeShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.feature.theme.toggle.ToggleThemeShowcasesProcessor
-import kotli.template.multiplatform.compose.showcases.navigation.NavigationShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.dataflow.cache.CacheShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.dataflow.http.HttpShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.dataflow.keyvalue.KeyValueShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.dataflow.paging.PagingShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.loader.data.DataLoaderShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.passcode.PasscodeShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.theme.ThemeShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.theme.change.ChangeThemeShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.theme.toggle.ToggleThemeShowcasesProcessor
+import kotli.template.multiplatform.compose.showcases.userflow.navigation.NavigationShowcasesProcessor
 
 object ShowcasesProvider : BaseFeatureProvider() {
