+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/showcases/userflow/loader/data/DataLoaderShowcasesProcessor.kt
@@ -1,4 +1,4 @@
-package kotli.template.multiplatform.compose.showcases.feature.loader.data
+package kotli.template.multiplatform.compose.showcases.userflow.loader.data
 
 import kotli.engine.BaseFeatureProcessor
 import kotli.engine.TemplateState
