+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/showcases/userflow/theme/ThemeShowcasesProcessor.kt
@@ -1,4 +1,4 @@
-package kotli.template.multiplatform.compose.showcases.feature.theme
+package kotli.template.multiplatform.compose.showcases.userflow.theme
 
 import kotli.engine.BaseFeatureProcessor
 import kotli.engine.TemplateState
